feat: rename Toast to Toaster
