+++ example/src/App.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { Toast, toast } from "@steveyuowo/vue-hot-toast";
+import { Toaster, toast } from "@steveyuowo/vue-hot-toast";
 import "@steveyuowo/vue-hot-toast/vue-hot-toast.css";
 </script>
 
@@ -35,7 +35,7 @@ import "@steveyuowo/vue-hot-toast/vue-hot-toast.css";
       Error
     </button>
   </div>
-  <Toast />
+  <Toaster />
 </template>
 
 <style lang="scss" scoped>
